+++ src/commands/content-item/__mocks__/import.ts
@@ -2,12 +2,22 @@ import { ImportItemBuilderOptions } from '../../../interfaces/import-item-builde
 import { ConfigurationParameters } from '../../configure';
 import { Arguments } from 'yargs';
 
+type ReturnType = boolean | 'throw';
+let expectedReturn: ReturnType = true;
+
 export const calls: Arguments<ImportItemBuilderOptions & ConfigurationParameters>[] = [];
+export const setExpectedReturn = (value: ReturnType): void => {
+  expectedReturn = value;
+};
 
 export const handler = async (
   argv: Arguments<ImportItemBuilderOptions & ConfigurationParameters>
 ): Promise<boolean> => {
   calls.push(argv);
 
-  return true;
+  if (expectedReturn == 'throw') {
+    throw new Error('Forced throw in test.');
+  }
+
+  return expectedReturn;
 };