+++ src/commands/content-item/export.ts
@@ -137,6 +137,7 @@ const getContentItems = async (
     let newItems: ContentItem[];
     try {
       const allItems = await paginator(repository.related.contentItems.list, { status: 'ACTIVE' });
+
       Array.prototype.push.apply(repoItems, allItems);
       newItems = allItems.filter(item => item.folderId == null);
     } catch (e) {
@@ -339,8 +340,14 @@ export const handler = async (argv: Arguments<ExportItemBuilderOptions & Configu
     log.appendLine(resolvedPath);
     await ensureDirectoryExists(directory);
 
+    if (argv.exportedIds) {
+      argv.exportedIds.push(item.id as string);
+    }
+
     writeJsonToFile(resolvedPath, item);
   }
 
-  log.close();
+  if (typeof logFile !== 'object') {
+    await log.close();
+  }
 };