+++ src/interfaces/copy-item-builder-options.interface.ts
@@ -1,0 +1,34 @@
+import { CopyConfig } from '../common/content-item/copy-config';
+import { FileLog } from '../common/file-log';
+
+export interface CopyItemBuilderOptions {
+  srcRepo?: string;
+  srcFolder?: string;
+
+  dstRepo?: string;
+  dstFolder?: string;
+
+  dstHubId?: string;
+  dstClientId?: string;
+  dstSecret?: string;
+
+  schemaId?: string[] | string;
+  name?: string[] | string;
+
+  mapFile?: string;
+  force?: boolean;
+  validate?: boolean;
+  skipIncomplete?: boolean;
+  logFile?: string | FileLog;
+  copyConfig?: string | CopyConfig;
+
+  revertLog?: string;
+
+  lastPublish?: boolean;
+  publish?: boolean;
+  republish?: boolean;
+
+  excludeKeys?: boolean;
+
+  exportedIds?: string[];
+}